Add download_all
The argument download_all will download all albums of a user
into a file tree. Folders in which the albums reside on SmugMug
will be created locally. Since smugline depends on smugpy which
uses an old API version, folders are only recognized and made
up to a depth of two. Also, albums in the root of a user are
shown as if they are in the "Homepage/" folder.